juicebox_supplementary_files: Adds juicebox on all hic files.
