+++ src/encoded/static/components/browse/components/StackedBlockTable.js
@@ -2,13 +2,11 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Checkbox, Collapse } from 'react-bootstrap';
+import { Button, Checkbox, Collapse } from 'react-bootstrap';
 import _ from 'underscore';
 import { FacetList } from './FacetList';
 import { expFxn, Filters, console, isServerSide, analytics, object, Schemas } from './../../util';
 import { requestAnimationFrame } from './../../viz/utilities';
-
-
 
 /**
  * Label to show at top left of Name block.
@@ -64,7 +62,7 @@ export class StackedBlockNameLabel extends React.Component {
 export class StackedBlockName extends React.Component {
 
     static Label = StackedBlockNameLabel
-    
+
     static propTypes = {
         'columnClass' : PropTypes.string,
         'colWidthStyles' : PropTypes.object,
@@ -76,7 +74,7 @@ export class StackedBlockName extends React.Component {
         'style' : PropTypes.object,
         'visible' : PropTypes.bool, // ? forgot
         'verticalAlign' : PropTypes.string // CSS vertical-align property. Change alignment/positioning if wanted.
-        
+
     }
 
     static defaultProps = {
@@ -141,7 +139,7 @@ export class StackedBlockName extends React.Component {
                     addedProps.handleFileCheckboxChange = this.props.handleFileCheckboxChange;
                 }
             }
-            
+
             if (_.keys(addedProps).length > 0){
                 return React.cloneElement(c, addedProps, c.props.children);
             }
@@ -189,7 +187,7 @@ export class StackedBlockName extends React.Component {
  * Button to toggle collapse/visible of longer StacedkBlockLists. Used in StackedBlockLists.
  */
 export class StackedBlockListViewMoreButton extends React.Component {
-    
+
     static propTypes = {
         collapsibleChildren : PropTypes.array,
         collapsed : PropTypes.bool,
@@ -486,7 +484,7 @@ export class StackedBlock extends React.Component {
 
 
 export class FilePairBlock extends React.Component {
-    
+
     static accessionTriplesFromProps(props){
         var accessionTriples;
         try {
@@ -652,7 +650,7 @@ export class FileEntryBlockPairColumn extends React.Component {
         this.hasCheckbox = fileEntryBlockMixins.hasCheckbox.bind(this);
         this.renderCheckBox = fileEntryBlockMixins.renderCheckBox.bind(this);
     }
-    
+
     render(){
         var tableHasFilePairColumn = _.pluck(this.props.columnHeaders || [], 'title').indexOf('File Pair') > -1;
         return (
@@ -724,7 +722,7 @@ export class FileEntryBlock extends React.Component {
                 //return true;
             }),
             baseClassName = (className || '') + " col-file-detail item";
-        
+
         for (var i = 0; i < cols.length; i++){
 
             var col = cols[i],
@@ -853,9 +851,68 @@ export class FileEntryBlock extends React.Component {
         return <StackedBlock.Name.Label {...commonProperties} />;
     }
 
+    /**
+    * Add a link to an external site for some file types.
+    * @param file - Describes the file object that will be displayed.
+    */
+    renderExternalLink(){
+      var { file } = this.props;
+      console.log(file);
+      // Find out if the file format is hic.
+      // - It needs an external href link
+      //   - Either it needs a file format of 'hic'
+      //   - OR it has a detailed file type that contains 'hic'
+      var fileIsHic = (
+        file
+        && (
+          (file.file_format && file.file_format === 'hic')
+          || (
+            file.file_type_detailed && file.file_type_detailed.indexOf('(hic)') > -1)
+          )
+        && file.href
+      );
+
+      var externalLinkButton;
+      if (fileIsHic) {
+        // Make an external juicebox link.
+        var onClick = function(evt){
+          // If we're on the server side, there is no need to make an external link.
+          if (isServerSide()) return null;
+
+          // Get the href from the page.
+          var pageHref = file.href,
+              hrefParts = url.parse(pageHref),
+              host = hrefParts.protocol + '//' + hrefParts.host,
+              targetLocation = "http://aidenlab.org/juicebox/?hicUrl=" + host + file.href;
+
+          var win = window.open(targetLocation, '_blank');
+          win.focus();
+        }
+
+        // Build the juicebox button
+        externalLinkButton = (
+            <Button bsSize="xs" bsStyle="primary" className="text-600 inline-block clickable in-stacked-table-button" data-tip="Visualize this file in JuiceBox" onClick={onClick}>
+                <i className="icon icon-fw icon-external-link text-smaller"/>
+            </Button>
+        );
+      }
+
+      // Return the External link.
+      return externalLinkButton;
+    }
+
     renderName(){
         var { file, colWidthStyles } = this.props;
-        return <div className={"name col-file" + (file && file.accession ? ' mono-text' : '')} style={colWidthStyles ? colWidthStyles.file : null} children={[this.renderLabel(), this.renderCheckBox(), this.renderNameInnerTitle()]}/>;
+        return <div
+          className={
+            "name col-file" + (
+              file && file.accession ? ' mono-text' : ''
+            )}
+          style={colWidthStyles ? colWidthStyles.file : null}
+          children={[
+            this.renderLabel(), this.renderCheckBox(), this.renderNameInnerTitle(), this.renderExternalLink()
+          ]}
+        />;
     }
 
     render(){
@@ -946,7 +1003,7 @@ export class StackedBlockTable extends React.Component {
         delete this.lastColumnWidths;
         delete this.cache.origColumnWidths;
     }
-    
+
     totalColumnsWidth(origColumnWidths = this.cache.origColumnWidths){
         return _.reduce(origColumnWidths, function(m,v){ return m + v; }, 0);
     }
@@ -1015,7 +1072,7 @@ export class StackedBlockTable extends React.Component {
 
     /**
      * If we have a SelectedFilesController up the parent/ancestor chain that feeds us selectFile, selectedFiles, and unselectFile, this is the handler to use for checkbox stacked blocks.
-     * 
+     *
      * @param {string|string[]} uuid - String or list of strings (File Item UUID)
      * @param {Object|Object[]} fileObj - File Item JSON
      * @returns {void} - Nothing.
@@ -1072,7 +1129,7 @@ export class StackedBlockTable extends React.Component {
             if (typeof this.props.collapseLongLists === 'boolean' && typeof c.props.collapseLongLists !== 'boolean'){
                 addedProps.collapseLongLists = this.props.collapseLongLists;
             }
-            
+
             if (Object.keys(addedProps).length > 0){
                 return React.cloneElement(c, addedProps, c.props.children);
             }
@@ -1081,7 +1138,7 @@ export class StackedBlockTable extends React.Component {
     }
 
     render(){
-        
+
         // Cache for each render.
         var minTotalWidth = Math.max(this.props.width || 0, this.totalColumnsWidth(this.getOriginalColumnWidths()));
         this.lastColumnWidths = this.getColumnWidths();
